fixed production base path issue
