+++ src/constants/routes.ts
@@ -1,3 +1,3 @@
 import { routePaths } from '../routes/routePaths';
 export const loggedOutRoute = routePaths.login;
-export const loggedInRoute = routePaths.home(':string');
+export const loggedInRoute = routePaths.home;
